KT 1 ja 2 ylesanne tehtud

+++ bootKT/src/main/java/johan/Funktsioon.java
@@ -1,3 +1,3 @@
 public interface Funktsioon{
-	public int f(int kordaja);
+	public double f(double kordaja);
 }
